fix(test): repair stale import paths missed in colocation refactor
Two test files were merged into main on PR #59 after the colocation
refactor was originally drafted, so they kept their pre-refactor import
paths and broke unit tests on this branch:

- commands/apps/list.test.ts referenced ../../test/stubs.ts
- test/integration/completion.test.ts referenced ../lib/setup.ts

Both now match the convention used by every other file in their
respective directories.
